change arena design to pointer
